[REFACTOR] : 폴더명 변경
